Update and rename index.php to singleData.php
This is JS and PHP... The PHP pulls a single interval of data from a weather station database. The data is then used in HTML to show weather data at http://54.81.106.102. The JS converts the pressure PHP array to a JS array and compares in to known ranges of barometric pressure to render a correct weather icon.